refactor: 자신의 댓글이 아니면 수정 또는 삭제가 불가능하도록 기능 변경
화면에서 어떻게 처리해야할지는 고민을 좀 더 해봐야할 것 같습니다.

+++ src/main/java/com/codessquad/qna/question/AnswerController.java
@@ -24,7 +24,7 @@ public class AnswerController {
                                HttpSession session) {
         Object loginUserAttribute = session.getAttribute(CommonUtility.SESSION_LOGIN_USER);
         if (loginUserAttribute == null) {
-            return "redirect:/users/login";
+            return CommonUtility.REDIRECT_LOGIN_PAGE;
         }
 
         try {
@@ -46,12 +46,17 @@ public class AnswerController {
                                      Model model) {
         Object loginUserAttribute = session.getAttribute(CommonUtility.SESSION_LOGIN_USER);
         if (loginUserAttribute == null) {
-            return "redirect:/users/login";
+            return CommonUtility.REDIRECT_LOGIN_PAGE;
         }
 
         try {
-            Question question = questionRepository.findById(questionId).orElseThrow(() -> new NotFoundException("못찾음"));
+            User loginUser = (User) loginUserAttribute;
+            Question question = questionRepository.findById(questionId)
+                                                  .orElseThrow(() -> new NotFoundException("못찾음"));
             Answer answer = answerRepository.findByQuestionIdAndId(questionId, answerId);
+            if (!loginUser.equals(answer.getWriter())) {
+                return "redirect:/questions/" + questionId;
+            }
             model.addAttribute("question", question);
             model.addAttribute("answer", answer);
         } catch (NotFoundException e) {
@@ -68,10 +73,14 @@ public class AnswerController {
                                HttpSession session) {
         Object loginUserAttribute = session.getAttribute(CommonUtility.SESSION_LOGIN_USER);
         if (loginUserAttribute == null) {
-            return "redirect:/users/login";
+            return CommonUtility.REDIRECT_LOGIN_PAGE;
         }
 
+        User loginUser = (User) loginUserAttribute;
         Answer answer = answerRepository.findByQuestionIdAndId(questionId, answerId);
+        if (!loginUser.equals(answer.getWriter())) {
+            return "redirect:/questions/" + questionId;
+        }
         answer.setComment(comments);
         answer.setUpdatedDateTime(LocalDateTime.now());
         answerRepository.save(answer);
@@ -81,8 +90,18 @@ public class AnswerController {
 
     @DeleteMapping("/questions/{questionId}/answers/{answerId}")
     public String deleteAnswer(@PathVariable Long questionId,
-                               @PathVariable Long answerId) {
+                               @PathVariable Long answerId,
+                               HttpSession session) {
+        Object loginUserAttribute = session.getAttribute(CommonUtility.SESSION_LOGIN_USER);
+        if (loginUserAttribute == null) {
+            return CommonUtility.REDIRECT_LOGIN_PAGE;
+        }
+
+        User loginUser = (User) loginUserAttribute;
         Answer answer = answerRepository.findByQuestionIdAndId(questionId, answerId);
+        if (!loginUser.equals(answer.getWriter())) {
+            return "redirect:/questions/" + questionId;
+        }
         answerRepository.delete(answer);
         return "redirect:/questions/" + questionId;
     }
